Check in some optimization attempts
houghGPUv1.cu is our baseline, the previously named houghGPU.png
that was based on the CPU code. v2-v5 are attempted optimizations.
Please read comments on top of
each file to see what each file is based on and what optimizations were
attempted.

So far v3 is the fastest one.